requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-telegram-bot[webhooks]==21.3
supabase==2.5.1
python-dotenv==1.0.1
python-multipart==0.0.9
